requirements.txt
# Core dependencies
python-telegram-bot==20.7
apscheduler==3.10.4
openai==1.3.7
anthropic==0.7.8
telethon==1.32.1

# Database
sqlite3

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.10
scrapy==2.11.0

# Data processing
pandas==2.1.3
python-dateutil==2.8.2

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0

# Logging and monitoring
structlog==23.2.0

# Text processing
markdown==3.5.1
html2text==2020.1.16

# Async support
aiohttp==3.9.1
asyncio 